// src/modules/gift/dto/search-gift.dto.ts
export class SearchGiftDto {
  user_id?: string;
  gift_type?: string;
  start_time?: string;
  end_time?: string;
  date_time?: string;
  point?: string;
  level?: string;
  like?: string;
  comment?: string;
  course?: string;
  channel_id?: string;
  birth?: string;
  stock_qty?: string;
  search?: string;
  date?: string;
  coin?: string;
}